introduce pkg `v3_allocator`
Package v3_allocator implements a distributed allocation algorithm
building upon the `keyspace` and `push_relabel` packages. It defines
concepts of Items, Members, and Assignments - each of which may have
user-defined decoders - and an Allocator which will coordinate with
other instances to incrementally and maximually assign Items to
Members, with replication.

Issue #49