persister: quick-n-dirty EtcdV3 support
Minimally adapt Persister to make use of the EtcdV3 API and leases.
Re-enable TestPersistence, adapting it for use with an actual
embedded EtcdV3 server.

This change is not pretty, but allows Persister to work with either
EtcdV2 & EtcdV3. I'd like to broadly re-work Persister once we're
entirely on EtcdV3.

Issue #43